// src/api/generated/model/webhookCreationDtoHttpMethod.ts
/**
 * Generated by orval v7.20.0 🍺
 * Do not edit manually.
 * Cosy API
 * Management API for Cosy (Cost Optimised Server Yard).
 * OpenAPI spec version: v1.0
 */

export type WebhookCreationDtoHttpMethod = typeof WebhookCreationDtoHttpMethod[keyof typeof WebhookCreationDtoHttpMethod];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const WebhookCreationDtoHttpMethod = {
  GET: 'GET',
  POST: 'POST',
  PUT: 'PUT',
  DELETE: 'DELETE',
  PATCH: 'PATCH',
  HEAD: 'HEAD',
  OPTIONS: 'OPTIONS',
} as const;